requirements.txt
streamlit
langchain
langchain-community
langchain-openai
openai
pymupdf
python-dotenv
faiss-cpu
